backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1
PyJWT==2.3.0

#Tests
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0

#Tools
alembic==1.13.1